解决给gradio设置auth之后，gradio_client访问api失败的问题。
If auth is set to gradio, calling from gradio_client to api will fail with exception:
server rejected WebSocket connection: HTTP 403
Current gradio version in GPT-SoVITS is 3.38.0 and official fix is on 4.9.0 according to:
https://github.com/gradio-app/gradio/issues/6605
This commit upgrades gradio to latest version of 4.19.2 and modified deprecated arguments in code, and will fix server-side problem.
On the other hand, the client-side of gradio_client of the latest version of 0.10.1 raises a new exception when uploading a file to a api protected by auth:
ValueError: None
The fix for the problem has no effect to the server-side due to ongoing pull request:
https://github.com/gradio-app/gradio/pull/7555
so this commit may be useful.